% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{BayesLinReg}
\alias{BayesLinReg}
\title{Simulate from posterior distribution in linear regression}
\usage{
BayesLinReg(y, X, mu_0, Omega_0, v_0, sigma2_0, nIter)
}
\arguments{
\item{y}{n-by-1 vector with response data observations}

\item{X}{n-by-nCovs matrix with covariates, first column should be ones if you want an intercept.}

\item{mu_0}{prior mean for beta}

\item{Omega_0}{prior precision matrix for beta}

\item{v_0}{degrees of freedom in the prior for sigma2}

\item{sigma2_0}{location ("best guess") in the prior for sigma2}

\item{nIter}{Number of draws from the posterior}
}
\value{
list with a vector of sigma2 draws (sigma2Sample) and a matrix with beta draws (betaSample)
}
\description{
Simulate from the posterior distribution of the Gaussian linear regression model
\deqn{y = \mathbf{x}^\top\boldsymbol{\beta} + \varepsilon, \varepsilon \sim N(0,\sigma^2)}{y = x'*beta + eps, eps ~ N(0, sigma2)}
with conjugate prior
\deqn{\beta | \sigma^22 ~ N(\mu_0, \sigma^2Omega_0^{-1})}{beta | sigma2 ~ N(mu_0, sigma2*inv(Omega_0))}
\deqn{\sigma^2 \sim \mathrm{Inv-}\chi^2(v_0,\sigma^2_0)}{sigma2 ~ Inv-Chi2(v_0,sigma2_0)}
}
\examples{
library(bayeslearn)
X = cbind(1,cars$speed)
y = cars$speed
mu_0 = c(0,0); Omega_0 = 0.1*diag(2);v_0 = 5; sigma2_0 = 10;
postRes = BayesLinReg(y, X, mu_0, Omega_0, v_0, sigma2_0, nIter = 1000)
hist(postRes$sigma2Sample, breaks = 30, main = expression(paste("Posterior ", sigma^2)), xlab = "")
hist(postRes$betaSample[,1], breaks = 30, main = expression(paste("Posterior ", beta[0])), xlab = "")
hist(postRes$betaSample[,2], breaks = 30, main = expression(paste("Posterior ", beta[1])), xlab = "")
}
