% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{logPostLogisticReg}
\alias{logPostLogisticReg}
\title{Log posterior for logistic regression with normal prior.}
\usage{
logPostLogisticReg(beta_, y, X, mu, Sigma)
}
\arguments{
\item{beta_}{vector of regression coefficients}

\item{y}{n-vector with binary responses}

\item{X}{n-by-p matrix with covariates, including a column of ones for intercept}

\item{mu}{prior mean vector for the beta coefficients}

\item{Sigma}{prior covariance matrix for the beta coefficients}
}
\value{
log posterior density at beta.
}
\description{
\deqn{\mathrm{Pr}(y=1|\mathbf{x}) = \frac{1}{1 +\exp(-\mathbf{x}^\top\boldsymbol{\beta})}}{Pr(y = 1 | x) = 1/(1 + exp(-x*beta))}
with multivariate normal prior
\deqn{\beta ~ N(\mu, \Sigma}{beta ~ N(mu,Sigma)}
}
\examples{
library(bayeslearn)
y = SUdatasets::titanic$survived
X = cbind(1, SUdatasets::titanic$age)
logPostLogisticReg(c(0,1), y, X, mu = c(0,0), Sigma = 10^2*diag(2))
}
