% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rScaledInvChi2}
\alias{rScaledInvChi2}
\title{Simulates from the Scaled Inv-Chi2 distribution.}
\usage{
rScaledInvChi2(n, v, tau2)
}
\arguments{
\item{n}{the number of draws}

\item{v}{degrees of freedom parameter in the Scaled Inv-Chi2 distribution}

\item{tau2}{scale parameter in the Scaled Inv-Chi2 distribution}
}
\value{
vector with n draws
}
\description{
Simulates from the Scaled Inv-Chi2 distribution.
}
\examples{
library(bayeslearn)
draws = rScaledInvChi2(n = 10000, v = 5, tau2 = 10)
mean(draws) # Should be close to the theoretical (v/(v-2))*\tau2 = 16.667
}
